增加redis缓存功能

+++ src/main/java/cn/zjy/dayong/demo/controller/UserController.java
@@ -9,13 +9,17 @@ import com.baomidou.mybatisplus.plugins.Page;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ValueOperations;
+import org.springframework.data.redis.serializer.RedisSerializer;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 /**
  * <p>
@@ -36,6 +40,16 @@ public class UserController {
 	@Autowired
     private UserService userService;
 
+	@Autowired
+    private RedisTemplate redisTemplate;
+
+	//解决存入redis里的key没有序列化   \xac\xed\x00\x05t\x00\x06
+    @Autowired(required = false)
+    public void setRedisTemplate(RedisTemplate redisTemplate) {
+        RedisSerializer stringSerializer = new StringRedisSerializer();
+        redisTemplate.setKeySerializer(stringSerializer);
+        this.redisTemplate = redisTemplate;
+    }
     /**
      * 根据用户id查询用户   Restful 风格
      * 请求路径 : http://localhost:2080/user/getUserById/3
@@ -45,9 +59,20 @@ public class UserController {
 	@RequestMapping("/getUserById/{id}")
     @ResponseBody
 	public ResponseMessage getUserById(@PathVariable Integer id){
+        final String key = "user_" + id;
+        ValueOperations<String, User> operations = redisTemplate.opsForValue();
+        // 缓存存在
+        boolean hasKey = redisTemplate.hasKey(key);
+        if (hasKey) {
+            User user = operations.get(key);
+            logger.info("从缓存中获取了用户:{} ", user.toString());
+            return new ResponseMessage().ok().put("从redis缓存中获取用户成功!",user);
+        }
         User user = userService.selectById(id);
-        logger.info("获取到的用户为:{}",user);
-        return new ResponseMessage().ok().put("获取用户成功!",user);
+        //key:是redis里的key  user:是存放的对象   60:是缓存对象,默认时间是秒    TimeUnit.SECONDS:?
+        operations.set(key, user, 60, TimeUnit.SECONDS);
+        logger.info("从数据库中获取到的用户对象,插入缓存:{}", user.toString());
+        return new ResponseMessage().ok().put("从数据库中获取用户成功!",user.toString());
     }
 
     /**
@@ -73,9 +98,21 @@ public class UserController {
     @RequestMapping("/selectUserPage/{current}/{size}")
     @ResponseBody
     public ResponseMessage selectUserPage(@PathVariable Integer current, @PathVariable Integer size){
+
+        final String key = "pageUser_" + current + "_" + size;
+        ValueOperations<String, List<User>> operations = redisTemplate.opsForValue();
+        // 缓存存在
+        boolean hasKey = redisTemplate.hasKey(key);
+        if (hasKey) {
+            List<User> userList = operations.get(key);
+            logger.info("从缓存中获取了用户:{} ", userList);
+            return new ResponseMessage().ok().put("从redis缓存中获取用户集合成功!",userList);
+        }
+        //key:是redis里的key  user:是存放的对象   60:是缓存对象,默认时间是秒    TimeUnit.SECONDS:?
         List<User> userList = userService.selectPage(new Page<User>(current,size)).getRecords();
-        logger.info("获取到的分页用户为:{}",userList );
-        return new ResponseMessage().ok().put("分页查询用户成功!",userList);
+        operations.set(key, userList, 60, TimeUnit.SECONDS);
+        logger.info("从数据库中获取到的用户对象,插入缓存:{}", userList);
+        return new ResponseMessage().ok().put("从数据库中获取到分页查询用户成功!",userList);
     }
 
     /**
@@ -85,6 +122,15 @@ public class UserController {
     @RequestMapping("/deleteUser/{id}")
     @ResponseBody
     public ResponseMessage deleteUser(@PathVariable Integer id){
+        final String key = "pagetUser_" + id;
+        ValueOperations<String, User> operations = redisTemplate.opsForValue();
+        // 缓存存在
+        boolean hasKey = redisTemplate.hasKey(key);
+        if (hasKey) {
+            User user = operations.get(key);
+            logger.info("从缓存中删除了用户:{} ",user);
+            return new ResponseMessage().ok().put("从redis缓存中删除用户成功!",user);
+        }
         User user = userService.selectById(id);
         logger.info("查询到的用户为:{}",user );
         boolean b = userService.deleteById(Integer.valueOf(id).longValue());
@@ -116,13 +162,23 @@ public class UserController {
     @RequestMapping("/selectUser")
     @ResponseBody
     public ResponseMessage selectUser(){
+        final String key = "pageUser_";
+        ValueOperations<String, List<User>> operations = redisTemplate.opsForValue();
+        // 缓存存在
+        boolean hasKey = redisTemplate.hasKey(key);
+        if (hasKey) {
+            List<User> userList = operations.get(key);
+            logger.info("从缓存中获取了用户:{} ", userList);
+            return new ResponseMessage().ok().put("从redis缓存中获取用户集合成功!",userList);
+        }
         //new Page<User>(3, 2)  分页显示数据  2代表显示几条数据   3代表要显示的数据是从(3-1)*2 + 1条开始显示的
         List<User> userList = userService.selectPage(new Page<User>(2, 4),
                 new EntityWrapper<User>().eq("name", "李李李李")
                         .eq("sex", "男")
                         .between("age", "10", "50")).getRecords();
-        logger.info("多条件分页查询到的用户为:{}",userList );
-        return new ResponseMessage().ok().put("多条件分页查询到的用户为!",userList);
+        operations.set(key, userList, 60, TimeUnit.SECONDS);
+        logger.info("从数据库中获取到的用户对象,插入缓存:{}", userList);
+        return new ResponseMessage().ok().put("从数据库中获取到多条件分页查询用户成功!",userList);
     }
 
     /**
@@ -132,8 +188,24 @@ public class UserController {
     @RequestMapping("/selectOneUser/{id}")
     @ResponseBody
     public ResponseMessage selectOneUserById(@PathVariable Integer id){
+        final String key = "user_" + id;
+        ValueOperations<String, User> operations = redisTemplate.opsForValue();
+        // 缓存存在
+        boolean hasKey = redisTemplate.hasKey(key);
+        if (hasKey) {
+            User user = operations.get(key);
+            logger.info("原生mybatis从缓存中获取了用户:{} ", user.toString());
+            return new ResponseMessage().ok().put("使用原生mybatis从redis缓存中获取用户成功!",user);
+        }
         User user = userService.selectOneUserById(id);
+        operations.set(key, user, 60, TimeUnit.SECONDS);
         logger.info("原生mybatis查询到的用户为:{}",user );
-        return new ResponseMessage().ok().put("原生mybatis查询到的用户为!",user);
+        return new ResponseMessage().ok().put("原生mybatis从数据库中查询到的用户为!",user);
     }
+
+
+
+    /**
+     * springboot整合redis
+     */
 }
